+++ core/paginate.py
@@ -1,0 +1,279 @@
+from __future__ import annotations
+
+import asyncio
+from typing import Any, Dict, Optional, TYPE_CHECKING
+
+import discord
+from discord.ext import menus
+
+if TYPE_CHECKING:
+    from main import Timmy
+
+
+class RoboPages(discord.ui.View):
+    def __init__(
+            self,
+            source: menus.PageSource,
+            *,
+            interaction: discord.Interaction,
+            check_embeds: bool = True,
+            compact: bool = False,
+            bot: Timmy,
+    ):
+        super().__init__()
+        self.source: menus.PageSource = source
+        self.check_embeds: bool = check_embeds
+        self.interaction: discord.Interaction = interaction
+        self.message: Optional[discord.Message] = None
+        self.bot = bot
+        self.current_page: int = 0
+        self.compact: bool = compact
+        self.input_lock = asyncio.Lock()
+        self.clear_items()
+        self.fill_items()
+
+    def fill_items(self) -> None:
+        if not self.compact:
+            self.numbered_page.row = 1
+            self.stop_pages.row = 1
+
+        if self.source.is_paginating():
+            max_pages = self.source.get_max_pages()
+            use_last_and_first = max_pages is not None and max_pages >= 2
+            if use_last_and_first:
+                self.add_item(self.go_to_first_page)  # type: ignore
+            self.add_item(self.go_to_previous_page)  # type: ignore
+            if not self.compact:
+                self.add_item(self.go_to_current_page)  # type: ignore
+            self.add_item(self.go_to_next_page)  # type: ignore
+            if use_last_and_first:
+                self.add_item(self.go_to_last_page)  # type: ignore
+            if not self.compact:
+                self.add_item(self.numbered_page)  # type: ignore
+            self.add_item(self.stop_pages)  # type: ignore
+
+    async def _get_kwargs_from_page(self, page: int) -> Dict[str, Any]:
+        value = await discord.utils.maybe_coroutine(self.source.format_page, self, page)
+        if isinstance(value, dict):
+            return value
+        elif isinstance(value, str):
+            return {"content": value, "embed": None}
+        elif isinstance(value, discord.Embed):
+            return {"embed": value, "content": None}
+        else:
+            return {}
+
+    async def show_page(
+            self, interaction: discord.Interaction, page_number: int
+    ) -> None:
+        page = await self.source.get_page(page_number)
+        self.current_page = page_number
+        kwargs = await self._get_kwargs_from_page(page)
+        self._update_labels(page_number)
+        if kwargs:
+            if interaction.response.is_done():
+                if self.message:
+                    await self.message.edit(**kwargs, view=self)
+            else:
+                await interaction.response.edit_message(**kwargs, view=self)
+
+    def _update_labels(self, page_number: int) -> None:
+        self.go_to_first_page.disabled = page_number == 0
+        if self.compact:
+            max_pages = self.source.get_max_pages()
+            self.go_to_last_page.disabled = (
+                    max_pages is None or (page_number + 1) >= max_pages
+            )
+            self.go_to_next_page.disabled = (
+                    max_pages is not None and (page_number + 1) >= max_pages
+            )
+            self.go_to_previous_page.disabled = page_number == 0
+            return
+
+        self.go_to_current_page.label = str(page_number + 1)
+        self.go_to_previous_page.label = str(page_number)
+        self.go_to_next_page.label = str(page_number + 2)
+        self.go_to_next_page.disabled = False
+        self.go_to_previous_page.disabled = False
+        self.go_to_first_page.disabled = False
+
+        max_pages = self.source.get_max_pages()
+        if max_pages is not None:
+            self.go_to_last_page.disabled = (page_number + 1) >= max_pages
+            if (page_number + 1) >= max_pages:
+                self.go_to_next_page.disabled = True
+                self.go_to_next_page.label = "…"
+            if page_number == 0:
+                self.go_to_previous_page.disabled = True
+                self.go_to_previous_page.label = "…"
+
+    async def show_checked_page(
+            self, interaction: discord.Interaction, page_number: int
+    ) -> None:
+        max_pages = self.source.get_max_pages()
+        try:
+            if max_pages is None:
+                # If it doesn't give maximum pages, it cannot be checked
+                await self.show_page(interaction, page_number)
+            elif max_pages > page_number >= 0:
+                await self.show_page(interaction, page_number)
+        except IndexError:
+            # An error happened that can be handled, so ignore it.
+            pass
+
+    async def interaction_check(self, interaction: discord.Interaction) -> bool:
+        if interaction.user.id in (self.bot.owner_id, interaction.user.id):
+            return True
+        await interaction.response.send_message(
+            "This pagination menu cannot be controlled by you, sorry!", ephemeral=True
+        )
+        return False
+
+    async def on_timeout(self) -> None:
+        if self.message:
+            await self.message.edit(view=None)
+
+    #     async def on_error(self, error: Exception, item: discord.ui.Item, interaction: discord.Interaction) -> None:
+    #         if interaction.response.is_done():
+    #             await interaction.followup.send(f'An unknown error occurred, sorry {error}', ephemeral=True)
+    #             print(error)
+    #         else:
+    #             await interaction.response.send_message(f'An unknown error occurred, sorry {error}', ephemeral=True)
+    #             print(error)
+
+    async def start(self) -> None:
+        if (
+                self.check_embeds
+                and not self.interaction.channel.permissions_for(
+            self.interaction.guild.me
+        ).embed_links
+        ):
+            await self.interaction.response.send_message(
+                "Bot does not have embed links permission in this channel."
+            )
+            return
+
+        await self.source._prepare_once()
+        page = await self.source.get_page(0)
+        kwargs = await self._get_kwargs_from_page(page)
+        self._update_labels(0)
+        self.message = await self.interaction.response.send_message(**kwargs, view=self)
+
+    @discord.ui.button(label="≪", style=discord.ButtonStyle.grey)
+    async def go_to_first_page(
+            self,
+            interaction: discord.Interaction,
+            button: discord.ui.Button,
+    ):
+        """go to the first page"""
+        await self.show_page(interaction, 0)
+
+    @discord.ui.button(label="Back", style=discord.ButtonStyle.blurple)
+    async def go_to_previous_page(
+            self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
+        """go to the previous page"""
+        await self.show_checked_page(interaction, self.current_page - 1)
+
+    @discord.ui.button(label="Current", style=discord.ButtonStyle.grey, disabled=True)
+    async def go_to_current_page(
+            self, interaction: discord.Interaction, button: discord.ui.Button
+    ):
+        pass
+
+    @discord.ui.button(label="Next", style=discord.ButtonStyle.blurple)
+    async def go_to_next_page(
+            self,
+            interaction: discord.Interaction,
+            button: discord.ui.Button,
+    ):
+        """go to the next page"""
+        await self.show_checked_page(interaction, self.current_page + 1)
+
+    @discord.ui.button(label="≫", style=discord.ButtonStyle.grey)
+    async def go_to_last_page(
+            self,
+            interaction: discord.Interaction,
+            button: discord.ui.Button,
+    ):
+        """go to the last page"""
+        # The call here is safe because it's guarded by skip_if
+        await self.show_page(interaction, self.source.get_max_pages() - 1)
+
+    @discord.ui.button(label="Skip to page...", style=discord.ButtonStyle.grey)
+    async def numbered_page(
+            self,
+            interaction: discord.Interaction,
+            button: discord.ui.Button,
+    ):
+        """lets you type a page number to go to"""
+        if self.input_lock.locked():
+            await interaction.response.send_message(
+                "Already waiting for your response...", ephemeral=True
+            )
+            return
+
+        if self.message is None:
+            return
+
+        async with self.input_lock:
+            channel = self.message.channel
+            author_id = interaction.user and interaction.user.id
+            await interaction.response.send_message(
+                "What page do you want to go to?", ephemeral=True
+            )
+
+            def message_check(m):
+                return (
+                        m.author.id == author_id
+                        and channel == m.channel
+                        and m.content.isdigit()
+                )
+
+            try:
+                msg = await self.bot.wait_for(
+                    "message", check=message_check, timeout=30.0
+                )
+            except asyncio.TimeoutError:
+                await interaction.followup.send("Took too long.", ephemeral=True)
+                await asyncio.sleep(5)
+            else:
+                page = int(msg.content)
+                await msg.delete()
+                await self.show_checked_page(interaction, page - 1)
+
+    @discord.ui.button(label="Quit", style=discord.ButtonStyle.red)
+    async def stop_pages(
+            self,
+            interaction: discord.Interaction,
+            button: discord.ui.Button,
+    ):
+        """stops the pagination session."""
+        await interaction.response.defer()
+        await interaction.delete_original_message()
+        self.stop()
+
+
+class FieldPageSource(menus.ListPageSource):
+    """A page source that requires (field_name, field_value) tuple items."""
+
+    def __init__(self, entries, *, per_page=12):
+        super().__init__(entries, per_page=per_page)
+        self.embed = discord.Embed(colour=discord.Colour.blurple())
+
+    async def format_page(self, menu, entries):
+        self.embed.clear_fields()
+        self.embed.description = None
+
+        for key, value in entries:
+            self.embed.add_field(name=key, value=value, inline=False)
+
+        maximum = self.get_max_pages()
+        if maximum > 1:
+            text = (
+                f"Page {menu.current_page + 1}/{maximum} ({len(self.entries)} entries)"
+            )
+            self.embed.set_footer(text=text)
+
+        return self.embed
+
